+++ packages/app/src/renderer/src/apis/sync.ts
@@ -1,21 +1,32 @@
 import request from "./request";
 
+import type { SyncType } from "@biliLive-tools/types";
+
 const syncTestUpload = async (data: {
   remoteFolder: string;
-  execPath: string;
-  type: "baiduPCS" | "aliyunpan";
+  type: SyncType;
+  execPath?: string;
+  apiUrl?: string;
+  username?: string;
+  password?: string;
 }) => {
   const res = await request.post(`/sync/uploadTest`, data);
   return res.data;
 };
 
-const syncTestLogin = async (data: { execPath: string; type: "baiduPCS" | "aliyunpan" }) => {
+const syncTestLogin = async (data: {
+  execPath?: string;
+  type: SyncType;
+  apiUrl?: string;
+  username?: string;
+  password?: string;
+}) => {
   const res = await request.get(`/sync/isLogin`, { params: data });
   return res.data;
 };
 
-const loginByCookie = async (data: { cookie: string; execPath: string }) => {
-  const res = await request.post(`/sync/loginByCookie`, data);
+const baiduPCSLogin = async (data: { cookie: string; execPath: string }) => {
+  const res = await request.post(`/sync/baiduPCSLogin`, data);
   return res.data;
 };
 
@@ -29,7 +40,8 @@ const aliyunpanLogin = async (data: {
 
 const sync = async (data: {
   file: string;
-  type: "baiduPCS" | "aliyunpan";
+  type: SyncType;
+  targetPath: string;
   options: { removeOrigin: boolean };
 }) => {
   const res = await request.post(`/sync/sync`, data);
@@ -39,7 +51,7 @@ const sync = async (data: {
 export default {
   syncTestUpload,
   syncTestLogin,
-  loginByCookie,
+  baiduPCSLogin,
   aliyunpanLogin,
   sync,
 };